refactor: 이슈 목록 업데이트 함수 issuesController로 이동 및 함수명 변경
